casting for 64bit archs

+++ HelloMixpanel/HelloMixpanel/AppDelegate.m
@@ -109,7 +109,7 @@
 {
     self.bgTask = [application beginBackgroundTaskWithExpirationHandler:^{
 
-        NSLog(@"%@ background task %u cut short", self, self.bgTask);
+        NSLog(@"%@ background task %lu cut short", self, (unsigned long)self.bgTask);
 
         [application endBackgroundTask:self.bgTask];
         self.bgTask = UIBackgroundTaskInvalid;
@@ -117,7 +117,7 @@
 
     dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
 
-        NSLog(@"%@ starting background task %u", self, self.bgTask);
+        NSLog(@"%@ starting background task %lu", self, (unsigned long)self.bgTask);
 
         // track some events and set some people properties
         Mixpanel *mixpanel = [Mixpanel sharedInstance];
@@ -125,12 +125,12 @@
         [mixpanel track:@"Background Event"];
         [mixpanel.people set:@"Background Property" to:[NSDate date]];
 
-        NSLog(@"%@ ending background task %u", self, self.bgTask);
+        NSLog(@"%@ ending background task %lu", self, (unsigned long)self.bgTask);
         [application endBackgroundTask:self.bgTask];
         self.bgTask = UIBackgroundTaskInvalid;
     });
 
-    NSLog(@"%@ dispatched background task %u", self, self.bgTask);
+    NSLog(@"%@ dispatched background task %lu", self, (unsigned long)self.bgTask);
 
 }
